Add files via upload
Once you have your data in Tableau, use Tableau to re-create the analysis you did in Task 1, or find other ways to analyze the data using Tableau. You can use the “bin” functionality in Tableau to create ranges for different columns in the spreadsheet, for example, and also create calculated fields for simple kinds of data aggregation or combinations. For this task, the specific analysis you do is not nearly as important as practicing using the tool and creating dashboards with it.

Finally, use Tableau’s dashboard features to create a simple dashboard combining the analysis you have done to visualize on one screen the most important analyses of the data to show insights about qualified sales prospects. Keep your dashboard simple, with just 2- 4 charts that collectively tell the story of your data analysis.