Add files via upload
Your task is to draw a one-page process map for a simple version of a mortgage loan underwriting and origination process. Afterwards, consider opportunities for improving the process to enhance shareholder value and client experience using the analysis process. Perhaps a step could be done faster, automated, eliminated, or accomplished a different way with a new tool or a new JPMorgan Chase & Co. product or service offering. Think broadly and creatively. What opportunities for improvement do you see?